requirements.txt
ultralytics>=8.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
pillow>=10.0.0
openai>=1.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0